--- general/jsonrpc/go-client/profiles/test/client.yml
@@ -1,60 +1,0 @@
-# dubbo client yaml configure file
-
-check: true
-# client
-request_timeout : "3s"
-# connect timeout
-connect_timeout : "3s"
-
-# application config
-application:
-  organization : "ikurento.com"
-  name  : "BDTService"
-  module : "dubbogo user-info client"
-  version : "0.0.1"
-  owner : "ZX"
-  environment : "test"
-
-registries :
-  "hangzhouzk":
-    protocol: "zookeeper"
-    timeout	: "3s"
-    address: "127.0.0.1:2181"
-    username: ""
-    password: ""
-  "shanghaizk":
-    protocol: "zookeeper"
-    timeout	: "3s"
-    address: "127.0.0.1:2182"
-    username: ""
-    password: ""
-
-references:
-  "UserProvider":
-    # 可以指定多个registry，使用逗号隔开;不指定默认向所有注册中心注册
-    registry: "hangzhouzk"
-    protocol : "jsonrpc"
-    interface : "com.ikurento.user.UserProvider"
-    cluster: "failover"
-    methods :
-    - name: "GetUser"
-      retries: 3
-  "UserProvider1":
-    registry: "hangzhouzk"
-    protocol: "jsonrpc"
-    version : "2.0"
-    interface: "com.ikurento.user.UserProvider"
-    cluster: "failover"
-    methods:
-    - name: "GetUser"
-      retries: 3
-  "UserProvider2":
-    registry: "hangzhouzk"
-    protocol: "jsonrpc"
-    version : "2.0"
-    group: "as"
-    interface: "com.ikurento.user.UserProvider"
-    cluster: "failover"
-    methods:
-    - name: "GetUser"
-      retries: 3
