+++ general/jsonrpc/go-server/pkg/user_provider.go
@@ -1,4 +1,4 @@
-package main
+package pkg
 
 import (
 	"context"
